Made the change from int to int64_t because GTC sends so many packets that there is overflooding

+++ bandwidth_steering.h
@@ -16,20 +16,20 @@ struct optimization_parameters{
 	alglib::integer_1d_array ct; // basically tells the optimizer the i-th constrain has what inequality sign
 };
 
-void configure_constraints(int num_groups, 
+void configure_constraints(int64_t num_groups, 
 							std::string& C_str, 
 							std::string& d_str, 
 							optimization_parameters* opt_params,
-							std::vector<std::vector<float>>& traffic_matrix);
+							std::vector<std::vector<double>>& traffic_matrix);
 
-void configure_parameters(int num_groups, 
+void configure_parameters(int64_t num_groups, 
 							optimization_parameters* opt_params, 
-							std::vector< std::vector<float> >& traffic_matrix);
+							std::vector< std::vector<double> >& traffic_matrix);
 	
 
 
-void optimize_allocation(int num_groups, 
+void optimize_allocation(int64_t num_groups, 
 							optimization_parameters* opt_params, 
-							std::vector< std::vector<float> >& traffic_matrix);
+							std::vector< std::vector<double> >& traffic_matrix);
 
 };
